Update Nomulus GKE scaling
Reduce the requested resources. At ideal each pod consumes about 25m
CPU, so request 50m and set the utilization to 50% to make sure that
each pod is not stressed under load.

Update the min and max for each service. Frontend and pubapi serve
direct external requests, so we need to have more instances ready to go.